test: split slow zero-page chat tests
